Correct final version

+++ src/platform.cpp
@@ -1,21 +1,10 @@
 #include "../includes/platform.hpp"
 
-/**
- * @brief Constructeur de la classe Platform.
- *
- * @param screen_width La largeur de l'écran.
- * @param screen_height La hauteur de l'écran.
- */
 Platform::Platform(int screen_width, int screen_height)
     : screenWidth(screen_width), screenHeight(screen_height), posX(screen_width / 2 - PLATFORM_WIDTH / 2), posY(screen_height - 50), velX(0), smallPlatformMalus(false), malusTimer(0)
 {
 }
 
-/**
- * @brief Gère les événements de clavier pour déplacer la plateforme.
- *
- * @param e L'événement SDL.
- */
 void Platform::handleEvent(const SDL_Event &e)
 {
     if (e.type == SDL_KEYDOWN && e.key.repeat == 0)
@@ -44,11 +33,6 @@ void Platform::handleEvent(const SDL_Event &e)
     }
 }
 
-/**
- * @brief Déplace la plateforme en fonction du temps écoulé.
- *
- * @param deltaTime Le temps écoulé depuis la dernière mise à jour.
- */
 void Platform::move(float deltaTime)
 {
     posX += velX * deltaTime;
@@ -63,51 +47,33 @@ void Platform::move(float deltaTime)
     }
 }
 
-/**
- * @brief Affiche la plateforme sur l'écran.
- *
- * @param renderer Le renderer SDL.
- */
 void Platform::render(SDL_Renderer *renderer) const
 {
-    SDL_Rect fillRect = {static_cast<int>(posX), static_cast<int>(posY), PLATFORM_WIDTH, PLATFORM_HEIGHT};
+    SDL_Rect fillRect = {posX, posY, PLATFORM_WIDTH, PLATFORM_HEIGHT};
     if (smallPlatformMalus)
     {
-        fillRect.w /= 2;
+        fillRect.w /= 2; // Reduce width by half if malus is active
     }
-    SDL_SetRenderDrawColor(renderer, 0x00, 0x00, 0x00, 0xFF);
+    SDL_SetRenderDrawColor(renderer, 0x00, 0x00, 0x00, 0xFF); // Black color
     SDL_RenderFillRect(renderer, &fillRect);
 }
 
-/**
- * @brief Retourne le rectangle représentant la plateforme.
- *
- * @return Le rectangle SDL de la plateforme.
- */
 SDL_Rect Platform::getRect() const
 {
-    SDL_Rect rect = {static_cast<int>(posX), static_cast<int>(posY), PLATFORM_WIDTH, PLATFORM_HEIGHT};
+    SDL_Rect rect = {posX, posY, PLATFORM_WIDTH, PLATFORM_HEIGHT};
     if (smallPlatformMalus)
     {
-        rect.w /= 2;
+        rect.w /= 2; // Adjust width if malus is active
     }
     return rect;
 }
 
-/**
- * @brief Applique le malus de petite plateforme.
- */
 void Platform::applySmallPlatformMalus()
 {
     smallPlatformMalus = true;
-    malusTimer = 15.0f;
+    malusTimer = 15.0f; // Set the malus duration to 15 seconds
 }
 
-/**
- * @brief Met à jour l'état de la plateforme.
- *
- * @param deltaTime Le temps écoulé depuis la dernière mise à jour.
- */
 void Platform::update(float deltaTime)
 {
     if (smallPlatformMalus)
@@ -115,7 +81,7 @@ void Platform::update(float deltaTime)
         malusTimer -= deltaTime;
         if (malusTimer <= 0)
         {
-            smallPlatformMalus = false;
+            smallPlatformMalus = false; // Restore the platform size after the malus duration
         }
     }
 }
